feat(packages): migrate package endpoints to holdingsIQ v2 API (#687)
Closes: MODKBEKBJ-804

+++ src/main/java/org/folio/repository/providers/ProviderRepository.java
@@ -11,5 +11,5 @@ public interface ProviderRepository {
 
   CompletableFuture<Void> delete(String vendorId, UUID credentialsId, String tenantId);
 
-  CompletableFuture<List<Long>> findIdsByTagFilter(TagFilter tagFilter, UUID credentialsId, String tenantId);
+  CompletableFuture<List<Integer>> findIdsByTagFilter(TagFilter tagFilter, UUID credentialsId, String tenantId);
 }
